Fixing Color picker and other undefined console errors (#516)

+++ src/js/options.js
@@ -18,8 +18,14 @@ var restore = document.getElementById("restore");
 var voice;
 var localimage=document.getElementById("localupload");
 let themeColorPickerButton = document.getElementById("btn-color-theme");
-let themeColorPicker = new LdColorPicker(themeColorPickerButton);
-LdColorPicker.init();
+if(themeColorPickerButton) {
+    let themeColorPicker = new LdColorPicker(themeColorPickerButton);
+    LdColorPicker.init();
+    themeColorPicker.on("change", (color)=>{
+        themeColorPickerButton.style.backgroundColor = color;
+        theme = color;
+    });
+}
 
 
 const isImageURL = url => {
@@ -117,22 +123,21 @@ if(backgroundChange) {
     });
 }
 
-restore.addEventListener("click", ()=>{
-    if(confirm("This will remove your customized settings. This action is irreversible.")){
-        userItems = ["theValue", "voice", "msgTheme", "theme"];
-        userItems.forEach(item => {
-            localStorage.removeItem(item);
-        });
-    }
-});
+if(restore) {
+    restore.addEventListener("click", ()=>{
+        if(confirm("This will remove your customized settings. This action is irreversible.")){
+            userItems = ["theValue", "voice", "msgTheme", "theme"];
+            userItems.forEach(item => {
+                localStorage.removeItem(item);
+            });
+        }
+    });
+}
 
-localimage.addEventListener("change",()=>{
-    var tmppath = URL.createObjectURL(localimage.files[0]);
-    localStorage.setItem("theValue",tmppath);
-    alert("Image Updated");
-});
-
-themeColorPicker.on("change", function(color){
-    themeColorPickerButton.style.backgroundColor = color;
-    theme = color;
-});
+if(localimage) {
+    localimage.addEventListener("change",()=>{
+        var tmppath = URL.createObjectURL(localimage.files[0]);
+        localStorage.setItem("theValue",tmppath);
+        alert("Image Updated");
+    });
+}
